Added Referencing, Images, Tables and references.
Added figures and images, as well as cross reference (for figures and images) throughout the .md document. I also added the references for the .bib file.
I modified the functions plot_boxplot.py (added subtitles for each plot) and plot_hist.py (made the title line a comment, because when we create a figure it generates a title, then it would be misleading which title to follow).